pt2

+++ abstutil/src/clone.rs
@@ -32,7 +32,7 @@ impl Clone for Box<dyn Cloneable> {
 impl Cloneable for () {}
 impl Cloneable for bool {}
 impl Cloneable for usize {}
-impl Cloneable for f64 {}
+impl Cloneable for f32 {}
 impl Cloneable for String {}
 impl Cloneable for Option<String> {}
 impl Cloneable for (String, Box<dyn Cloneable>) {}
