+++ ezgui/src/tools/wizard.rs
@@ -186,12 +186,12 @@ impl<'a, 'b> WrappedWizard<'a, 'b> {
         )
     }
 
-    pub fn input_percent(&mut self, query: &str) -> Option<f64> {
+    pub fn input_percent(&mut self, query: &str) -> Option<f32> {
         self.input_something(
             query,
             None,
             Box::new(|line| {
-                line.parse::<f64>().ok().and_then(|num| {
+                line.parse::<f32>().ok().and_then(|num| {
                     if num >= 0.0 && num <= 1.0 {
                         Some(num)
                     } else {
